Added comments, tested on GNU/Linux, and more...
Comments added on all .vb files telling what it's for.

Most stuff works on GNU/Linux with Mono except things in QuickIO. See
the comments on there for more details.

Added 5 more threads (10 threads in total) to http and ping ddos attack
because it hardly made a dent in my server with the original 5.

More testing on GNU/Linux to come!